Added a measure to try fixing a crash when multiple deaths happen at the same time. Changed a bit how Sardine's bounties are sorted.

+++ Behaviors/CombatBehavior.cs
@@ -70,7 +70,7 @@ namespace terraguardians
                         SpecialWeaponSlot++;
                         if (SpecialWeaponSlot >= 10)
                         {
-                            SpecialWeaponSlot -= (byte)10;
+                            SpecialWeaponSlot -= 10;
                         }
                     }
                 }
